@@ -13,51 +13,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "github:angular/animations-builds",
-    "@angular/cdk": "github:angular/cdk-builds",
-    "@angular/common": "7.2.1",
-    "@angular/compiler": "7.2.1",
-    "@angular/core": "7.2.1",
-    "@angular/forms": "7.2.1",
-    "@angular/http": "7.2.1",
-    "@angular/material": "github:angular/material2-builds",
-    "@angular/platform-browser": "7.2.1",
-    "@angular/platform-browser-dynamic": "7.2.1",
-    "@angular/platform-server": "7.2.1",
-    "@angular/router": "7.2.1",
-    "@aspnet/signalr": "^1.1.0",
-    "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.5",
+    "@angular/animations": "^8.0.0-rc.2",
+    "@angular/cdk": "^8.0.0-rc.0",
+    "@angular/common": "8.0.0-rc.2",
+    "@angular/compiler": "8.0.0-rc.2",
+    "@angular/core": "8.0.0-rc.2",
+    "@angular/forms": "8.0.0-rc.2",
+    "@angular/material": "^8.0.0-rc.0",
+    "@angular/platform-browser": "8.0.0-rc.2",
+    "@angular/platform-browser-dynamic": "8.0.0-rc.2",
+    "@angular/platform-server": "8.0.0-rc.2",
+    "@angular/router": "8.0.0-rc.2",
+    "@aspnet/signalr": "^3.0.0-preview4-19216-03",
+    "@nguniversal/module-map-ngfactory-loader": "^8.0.0-rc.0",
     "angularx-social-login": "^1.2.6",
     "aspnet-prerendering": "^3.0.1",
-    "bootstrap": "^4.2.1",
-    "core-js": "^2.4.1",
+    "bootstrap": "^4.3.1",
+    "core-js": "^3.0.1",
     "lodash": "^4.17.11",
     "material-design-icons": "^3.0.1",
-    "rxjs": "^6.3.3",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.28"
+    "rxjs": "^6.5.1",
+    "tslib": "^1.9.3",
+    "zone.js": "~0.9.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.0",
-    "@angular/cli": "7.2.2",
-    "@angular/compiler-cli": "7.2.1",
-    "@angular/language-service": "7.2.1",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
+    "@angular-devkit/build-angular": "~0.800.0-rc.2",
+    "@angular/cli": "^8.0.0-rc.2",
+    "@angular/compiler-cli": "8.0.0-rc.2",
+    "@angular/language-service": "8.0.0-rc.2",
+    "@types/jasmine": "~3.3.12",
+    "@types/jasminewd2": "~2.0.6",
     "@types/lodash": "^4.14.120",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@types/node": "~11.13.8",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.1.0",
     "tslint": "~5.9.1",
-    "typescript": "3.2.4"
+    "typescript": "^3.4.5"
   },
   "optionalDependencies": {
     "node-sass": "^4.9.0"
