@@ -1,4 +1,4 @@
-import { SOCIAL_CONFIG } from '../../config';
+import { SOCIAL_CONFIG } from '@br/config';
 
 export function provideConfig() {
     return SOCIAL_CONFIG;
