@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { MatDialog } from '@angular/material';
+import { MatDialog } from "@angular/material/dialog";
 import { BookEditDialogComponent } from '../../../components';
 import { AuthorBooksService } from '../../../services';
 import { WebResult, Book, OperationResult } from '../../../models';
