@@ -1,8 +1,9 @@
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
+import { environment } from '@br/env/environment';
 import { AppModule } from './app/app.module';
-import { environment } from './environments/environment';
+
 
 export function getBaseUrl() {
   return document.getElementsByTagName('base')[0].href;
